separated data.py and mhd.py in pymiso
